+++ src/frontend/src/features/participants/components/ParticipantRow.tsx
@@ -5,7 +5,11 @@ import { Text } from '@/primitives/Text'
 import { useTranslation } from 'react-i18next'
 import { Avatar } from '@/components/Avatar'
 import { getParticipantColor } from '@/features/rooms/utils/getParticipantColor'
-import { getParticipantIsRoomOwner } from '@/features/rooms/utils/getParticipantIsRoomAdminOrOwner'
+import {
+  getParticipantIsRoomAdmin,
+  getParticipantIsRoomOwner,
+  getParticipantIsRoomMember,
+} from '@/features/rooms/utils/getParticipantIsRoomAdminOrOwner'
 import { type LocalParticipant, type Participant, Track } from 'livekit-client'
 import { isLocal } from '@/utils/livekit'
 import {
@@ -148,6 +152,12 @@ export const ParticipantRow = ({ participant }: ParticipantListItemProps) => {
           </Text>
           {getParticipantIsRoomOwner(participant) && (
             <Text variant="xsNote">{t('participants.host')}</Text>
+          )}
+          {getParticipantIsRoomAdmin(participant) && (
+            <Text variant="xsNote">{t('participants.cohost')}</Text>
+          )}
+          {getParticipantIsRoomMember(participant) && (
+            <Text variant="xsNote">{t('participants.member')}</Text>
           )}
         </VStack>
       </HStack>