+++ src/frontend/src/features/participants/components/menu/items/RemoveMenuItem.tsx
@@ -1,4 +1,4 @@
-import type { Participant } from 'livekit-client'
+import React from 'react'
 import { menuRecipe } from '@/primitives/menuRecipe'
 import { HStack } from '@/styled-system/jsx'
 import { RiCloseLine } from '@remixicon/react'
@@ -6,23 +6,30 @@ import { MenuItem } from 'react-aria-components'
 import { useRemoveParticipant } from '@/features/rooms/api/removeParticipant'
 import { useTranslation } from 'react-i18next'
 
-export const RemoveMenuItem = ({
-  participant,
-}: {
-  participant: Participant
-}) => {
-  const { t } = useTranslation('rooms', { keyPrefix: 'participantMenu.remove' })
-  const { removeParticipant } = useRemoveParticipant()
-  return (
-    <MenuItem
-      aria-label={t('ariaLabel', { name: participant.name })}
-      className={menuRecipe({ icon: true }).item}
-      onAction={() => removeParticipant(participant)}
-    >
-      <HStack gap={0.25}>
-        <RiCloseLine size={20} aria-hidden />
-        {t('label')}
-      </HStack>
-    </MenuItem>
-  )
+type RemoveMenuItemProps = {
+  identity: string
+  displayedName?: string
 }
+
+export const RemoveMenuItem = React.memo(
+  ({ identity, displayedName }: RemoveMenuItemProps) => {
+    const { t } = useTranslation('rooms', {
+      keyPrefix: 'participantMenu.remove',
+    })
+    const { removeParticipant } = useRemoveParticipant()
+    return (
+      <MenuItem
+        aria-label={t('ariaLabel', { name: displayedName || identity })}
+        className={menuRecipe({ icon: true }).item}
+        onAction={() => removeParticipant(identity)}
+      >
+        <HStack gap={0.25}>
+          <RiCloseLine size={20} aria-hidden />
+          {t('label')}
+        </HStack>
+      </MenuItem>
+    )
+  }
+)
+
+RemoveMenuItem.displayName = 'RemoveMenuItem'